requirements.txt
openai
chromadb
python-dotenv
fastapi
uvicorn
pydantic
streamlit
numpy
sentence-transformers